requirements.txt
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.3
coverage==6.3.1
decorator==5.1.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
executing==0.8.2
Faker==12.0.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mixer==7.2.1
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-ipdb @ git+git://github.com/mverteuil/pytest-ipdb.git@8e7f57b94e41832795c7171c999828b2999bea91
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
